+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 # next
 
-- skiplines option renamed to skip
-- headerlines option renamed to header
+- rename skiplines option to skip
+- rename headerlines option to header
+- move Path and subclasses out of vdtui
+- TextSheet source is any iterable of strings
+- Sheet.filetype provides default save filename extension
 
 ## 0.97 (2017-10-05)
 - Features